+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @packages/hooks
 
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @packages/utils@1.1.1
+  - @packages/tests@1.1.1
+  - @packages/context@1.1.1
+
 ## 1.0.2
 
 ### Patch Changes