+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @packages/ui
 
+## 1.1.1
+
 ## 1.1.0
 
 ### Minor Changes