+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @packages/utils
 
+## 1.1.1
+
+### Patch Changes
+
+- Fix bug
+
 ## 1.0.2
 
 ### Patch Changes